You Are An Idiot (Safe)
This is a SAFE Version of the YOU ARE AN IDIOT Old Virus aka Trojan OFFIZ / Trojan.JS.YouAreAnIdiot

This is Coded in python I Used a Converter to get it into EXE, You will need to READ The README.txt before anything! I am not responsible if you do not read it!